Updated Syntax that generates correct output for nested list and counts number of months correctly
Also had to remove the multiple print rows that I had added in throughout the code to simply verify the data and process.  Unfortunately, having all of the print(row) commands caused the final output of the list to have each row listed 3 - 4 or more times.  Also, had to move the pop(0) out of the loop/iteration.  Sent questions to tutor to determine to understand the pint(rows) issues with the output